+++ partner_firstname/tests/test_partner_firstname.py
@@ -25,46 +25,10 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-from openerp.tests.common import TransactionCase
-from .. import exceptions as ex
+from .base import BaseCase
 
 
-class PartnerFirstnameCase(TransactionCase):
-    def setUp(self):
-        super(PartnerFirstnameCase, self).setUp()
-        self.check_fields = True
-        self.expect(u"Núñez", u"Fernán")
-        self.create_original()
-
-    def create_original(self):
-        self.original = self.env["res.partner"].create({
-            "lastname": self.lastname,
-            "firstname": self.firstname})
-
-    def expect(self, lastname, firstname, name=None):
-        """Define what is expected in each field when ending."""
-        self.lastname = lastname
-        self.firstname = firstname
-        self.name = name or u"%s %s" % (lastname, firstname)
-
-    def tearDown(self):
-        if self.check_fields:
-            if not hasattr(self, "changed"):
-                self.changed = self.original
-
-            for field in ("name", "lastname", "firstname"):
-                self.assertEqual(
-                    getattr(self.changed, field),
-                    getattr(self, field),
-                    "Test failed with wrong %s" % field)
-
-        super(PartnerFirstnameCase, self).tearDown()
-
-    def test_copy(self):
-        """Copy the partner and compare the result."""
-        self.expect(self.lastname, u"%s (copy)" % self.firstname)
-        self.changed = self.original.with_context(lang="en_US").copy()
-
+class PartnerContactCase(BaseCase):
     def test_update_lastname(self):
         """Change lastname."""
         self.expect(u"newlästname", self.firstname)
@@ -83,14 +47,24 @@ class PartnerFirstnameCase(TransactionCase):
         # Need this to refresh the ``name`` field
         self.original.invalidate_cache()
 
-    def test_no_names(self):
-        """Test that you cannot set a partner/user without names."""
-        self.check_fields = False
-        with self.assertRaises(ex.EmptyNamesError):
-            self.original.firstname = self.original.lastname = False
+
+class PartnerCompanyCase(BaseCase):
+    def create_original(self):
+        super(PartnerCompanyCase, self).create_original()
+        self.original.is_company = True
+
+    def test_copy(self):
+        super(PartnerCompanyCase, self).test_copy()
+        self.expect(self.name, False, self.name)
+
+    def test_company_inverse(self):
+        """Test the inverse method in a company record."""
+        name = u"Thïs is a Companŷ"
+        self.expect(name, False, name)
+        self.original.name = name
 
 
-class UserFirstnameCase(PartnerFirstnameCase):
+class UserCase(PartnerContactCase):
     def create_original(self):
         self.original = self.env["res.users"].create({
             "name": u"%s %s" % (self.lastname, self.firstname),
